# api/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Medfetch API"
  },
  "components": {
    "schemas": {
      "TableResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "nullable": true
              }
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Filter": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "operator": {
            "type": "string",
            "enum": [
              "equals",
              "contains",
              "greaterThan",
              "lessThan",
              "between"
            ]
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    }
                  ]
                }
              }
            ]
          }
        },
        "required": [
          "field",
          "operator",
          "value"
        ]
      },
      "Sort": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        },
        "required": [
          "field",
          "direction"
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "minimum": 1,
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          }
        }
      },
      "TableOperation": {
        "type": "object",
        "properties": {
          "table": {
            "type": "string",
            "enum": [
              "Patient",
              "Procedure"
            ]
          },
          "operation": {
            "type": "string",
            "enum": [
              "select",
              "insert",
              "update",
              "delete"
            ]
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Filter"
            }
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sort"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "table",
          "operation"
        ]
      },
      "BulkResponse": {
        "type": "object",
        "properties": {
          "affectedRows": {
            "type": "number"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "affectedRows"
        ]
      },
      "BulkOperation": {
        "type": "object",
        "properties": {
          "table": {
            "type": "string",
            "enum": [
              "Patient",
              "Procedure"
            ]
          },
          "operation": {
            "type": "string",
            "enum": [
              "update",
              "delete"
            ]
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Filter"
            }
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "table",
          "operation",
          "filters"
        ]
      },
      "SchemaResponse": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "nullable": {
                  "type": "boolean"
                }
              },
              "required": [
                "name",
                "type",
                "nullable"
              ]
            }
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "schema"
        ]
      },
      "SchemaOperation": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "get",
              "addColumn",
              "removeColumn"
            ]
          },
          "table": {
            "type": "string",
            "enum": [
              "Patient",
              "Procedure"
            ]
          },
          "column": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "TEXT",
                  "INTEGER",
                  "REAL",
                  "BOOLEAN",
                  "DATE"
                ]
              },
              "nullable": {
                "type": "boolean",
                "default": true
              }
            },
            "required": [
              "name",
              "type"
            ]
          }
        },
        "required": [
          "operation",
          "table"
        ]
      },
      "SQL": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "sql": {
            "type": "string",
            "example": "SELECT * FROM Patient"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "summary"
        ]
      },
      "POST": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "example": "Get me all happy Patients"
          }
        },
        "required": [
          "query"
        ]
      },
      "CreateWorkspaceBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Workspace": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateWorkspaceBody"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true
              },
              "vfsType": {
                "type": "string",
                "enum": [
                  "opfs",
                  "kvfs"
                ]
              }
            },
            "required": [
              "id",
              "vfsType"
            ]
          }
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/table": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableOperation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Table operation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/bulk": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOperation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk operation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/schema": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaOperation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schema operation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/nl2sql": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SQL from natural language",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SQL"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201 Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "401 Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "409 Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "200 OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  }
                }
              }
            }
          },
          "401": {
            "description": "401 Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "400 Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401 Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404 Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}